+++ src/com/pvporbit/freetype/SizeMetrics.java
@@ -8,35 +8,35 @@ public class SizeMetrics extends Pointer {
 		super(pointer);
 	}
 
-	public long getAscender() {
+	public int getAscender() {
 		return FreeType.FT_Size_Metrics_Get_ascender(pointer);
 	}
 
-	public long getDescender() {
+	public int getDescender() {
 		return FreeType.FT_Size_Metrics_Get_descender(pointer);
 	}
 
-	public long getHeight() {
+	public int getHeight() {
 		return FreeType.FT_Size_Metrics_Get_height(pointer);
 	}
 
-	public long getMaxAdvance() {
+	public int getMaxAdvance() {
 		return FreeType.FT_Size_Metrics_Get_max_advance(pointer);
 	}
 
-	public long getXppem() {
+	public int getXppem() {
 		return FreeType.FT_Size_Metrics_Get_x_ppem(pointer);
 	}
 
-	public long getYppem() {
+	public int getYppem() {
 		return FreeType.FT_Size_Metrics_Get_y_ppem(pointer);
 	}
 
-	public long getXScale() {
+	public int getXScale() {
 		return FreeType.FT_Size_Metrics_Get_x_scale(pointer);
 	}
 
-	public long getYScale() {
+	public int getYScale() {
 		return FreeType.FT_Size_Metrics_Get_y_scale(pointer);
 	}
 }
